Tried to implement a camera function and a shuffleboard widget
The camera function we tried to build used a (likely) outdated source to build and thus is incomplete but in progress. The widget was going to be a menu to select an autonomous function to run. However we could not find reliable documentation that would give us an answer on how to create a drop down menu (Maybe use one button as a last resort? Not a fantastic solution but it would work.) and so there shouldn't be any remains of our work.